generators/wikimedia/requirements.txt
certifi==2018.11.29
chardet==3.0.4
CommonsDownloader==0.5.3
Django==1.11.20
docx==0.2.4
idna==2.8
image==1.5.27
lorem==0.1.1
lxml==4.3.2
mwclient==0.9.3
numpy==1.16.2
oauthlib==3.0.1
Pillow==5.4.1
python-slugify==3.0.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
